requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
Django==5.0
django-cogwheels==0.3
django-cors-headers==4.3.1
django-filter==23.5
django-modelcluster==6.1
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
dnspython==2.4.2
draftjs-exporter==2.1.7
email-validator==2.1.0.post1
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
pillow-heif==0.14.0
platformdirs==4.1.0
py3dns==4.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
validate-email==1.3
wagtail==5.2.2
wagtail-color-panel==1.5.0
wagtail-headless-preview==0.7.0
wagtailmenus==3.1.9
webencodings==0.5.1
Willow==1.6.3
